Split R 32-bit int generation into separate header.
